+++ compiler/parser/enrobage.cpp
@@ -395,7 +395,7 @@ const char* fileBasename(const char* name)
 string fileDirname(const string& name)
 {
     const char*        base = fileBasename(name.c_str());
-    const unsigned int size = (const unsigned int)(base - name.c_str());
+    const unsigned int size = base - name.c_str();
     string             dirname;
 
     if (size == 0) {
